BACKPORT: net: Add and use skb_list_del_init().
It documents what is happening, and eliminates the spurious list
pointer poisoning.

In the long term, in order to get proper list head debugging, we
might want to use the list poison value as the indicator that
an SKB is a singleton and not on a list.

Change-Id: I591ef9b88b4a36f6b7dea788dfb9c34f739884f7
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 992cba7e276d438ac8b0a8c17b147b37c8c286f7)